requirements.txt
anyio==3.6.1
certifi==2022.9.14
cffi==1.15.1
click==8.1.3
cryptography==38.0.1
fastapi==0.85.0
h11==0.12.0
httpcore==0.15.0
httptools==0.5.0
httpx==0.23.0
idna==3.4
pycparser==2.21
pydantic==1.10.2
PyJWT==2.5.0
python-dotenv==0.21.0
PyYAML==6.0
rfc3986==1.5.0
sniffio==1.3.0
starlette==0.20.4
types-cryptography==3.3.23
typing_extensions==4.3.0
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.17.0
websockets==10.3
